requirements.txt
pyyaml>=6.0
pytz>=2023.3
ics>=0.8.0
python-dateutil>=2.8.2
click>=8.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.5.0
python-multipart>=0.0.6

